requirements.txt
google-genai>=1.0.0
mlflow[genai]>=2.18.0
python-dotenv>=1.0.0
pydantic>=2.0.0
PyPDF2>=3.0.0
bert-score>=0.3.13
rouge-score>=0.1.2
